+++ src/schemathesis/specs/graphql/loaders.py
@@ -35,6 +35,7 @@ def from_path(
     code_sample_style: str = CodeSampleStyle.default().name,
     rate_limit: Optional[str] = None,
     encoding: str = "utf8",
+    sanitize_output: bool = True,
 ) -> GraphQLSchema:
     """Load GraphQL schema via a file from an OS path.
 
@@ -50,6 +51,7 @@ def from_path(
             code_sample_style=code_sample_style,
             location=pathlib.Path(path).absolute().as_uri(),
             rate_limit=rate_limit,
+            sanitize_output=sanitize_output,
         )
 
 
@@ -77,6 +79,7 @@ def from_url(
     code_sample_style: str = CodeSampleStyle.default().name,
     wait_for_schema: Optional[float] = None,
     rate_limit: Optional[str] = None,
+    sanitize_output: bool = True,
     **kwargs: Any,
 ) -> GraphQLSchema:
     """Load GraphQL schema from the network.
@@ -119,6 +122,7 @@ def from_url(
         data_generation_methods=data_generation_methods,
         code_sample_style=code_sample_style,
         rate_limit=rate_limit,
+        sanitize_output=sanitize_output,
     )
 
 
@@ -131,6 +135,7 @@ def from_file(
     code_sample_style: str = CodeSampleStyle.default().name,
     location: Optional[str] = None,
     rate_limit: Optional[str] = None,
+    sanitize_output: bool = True,
 ) -> GraphQLSchema:
     """Load GraphQL schema from a file descriptor or a string.
 
@@ -157,6 +162,7 @@ def from_file(
         code_sample_style=code_sample_style,
         location=location,
         rate_limit=rate_limit,
+        sanitize_output=sanitize_output,
     )
 
 
@@ -169,6 +175,7 @@ def from_dict(
     data_generation_methods: DataGenerationMethodInput = DEFAULT_DATA_GENERATION_METHODS,
     code_sample_style: str = CodeSampleStyle.default().name,
     rate_limit: Optional[str] = None,
+    sanitize_output: bool = True,
 ) -> GraphQLSchema:
     """Load GraphQL schema from a Python dictionary.
 
@@ -194,6 +201,7 @@ def from_dict(
         data_generation_methods=prepare_data_generation_methods(data_generation_methods),
         code_sample_style=_code_sample_style,
         rate_limiter=rate_limiter,
+        sanitize_output=sanitize_output,
     )  # type: ignore
     dispatch("after_load_schema", hook_context, instance)
     return instance
@@ -207,6 +215,7 @@ def from_wsgi(
     data_generation_methods: DataGenerationMethodInput = DEFAULT_DATA_GENERATION_METHODS,
     code_sample_style: str = CodeSampleStyle.default().name,
     rate_limit: Optional[str] = None,
+    sanitize_output: bool = True,
     **kwargs: Any,
 ) -> GraphQLSchema:
     """Load GraphQL schema from a WSGI app.
@@ -230,6 +239,7 @@ def from_wsgi(
         data_generation_methods=data_generation_methods,
         code_sample_style=code_sample_style,
         rate_limit=rate_limit,
+        sanitize_output=sanitize_output,
     )
 
 
@@ -241,6 +251,7 @@ def from_asgi(
     data_generation_methods: DataGenerationMethodInput = DEFAULT_DATA_GENERATION_METHODS,
     code_sample_style: str = CodeSampleStyle.default().name,
     rate_limit: Optional[str] = None,
+    sanitize_output: bool = True,
     **kwargs: Any,
 ) -> GraphQLSchema:
     """Load GraphQL schema from an ASGI app.
@@ -263,6 +274,7 @@ def from_asgi(
         data_generation_methods=data_generation_methods,
         code_sample_style=code_sample_style,
         rate_limit=rate_limit,
+        sanitize_output=sanitize_output,
     )